Change representation of query string parameters
Using a dict for query string parameters is problematic for a couple of
reasons: query strings are ordered whereas the enumeration order of a
dict's keys is undefined, a query string may have a key without a value,
and a query string may contain multiple occurrences of the same key.

Because of these reasons, I changed the representation for the query
string to be an array of (String, String?) tuples (typealiased to
"QueryParameter"), and added some QueryParameter convenience functions.