StorageBlobClient API review and documentation (#236)
* Move ResumableOperation to StorageBlob library

* Merge ResumableTransfer into ResumableOperation

* Revise TransferDelegate interface

Simplified based on what a consumer actually cares about

* Add a UUID "id" property to all CoreData transfer models

Allows retrieving an individual transfer by its UUID later

* Remove unknown in CoreData enum values

We are managing the backing Int16, so force-unwrapping here is safe and
it's better not to expose unknown to the user.

* Promote the transferDelegate out of the StorageBlobClientOptions

It makes sense for developers to attach the delegate directly to the
StorageBlobClient. Additionally it's easier for developers to detach the
delegate later.

* Create separate StorageBlobClient initializers for different creds types

* Clean up public API for StorageBlobClient

* Uplift TransferManager operations to StorageBlobClient

* Tighten access controls for AzureStorageBlob

This commit also removes classes that aren't needed in the current
implementation, and moves things around into more logical groups within
the .xcodeproj

* Remove duplicate AzureStorage CoreData model

* Adding API docstrings

* Update Podfile.lock

* Expose BlobTransfer in the public API

* Document the HTTPHeaders extension

* Fix compiler warnings.

* Add missing StorageHTTPHeader values

* Remove swiftlint file-level disables

This commit refactors some places that were causing fatal swiftlint
issues.

* Expose additional properties and methods needed by the demo app

These are properties / methods that we believe are generally useful, and
not just to explicitly support the demo app itself. The demo app is just
a good way of helping us to locate properties that customers may want to
access.

* Fix up the demo app, post API review