API review - followup PR (#239)
* Use URLs rather than Strings when building clients

Also try to be consistent with casing of parameter names, "URL" vs "Url"

* Always explicitly specify source and destination of transfers.

This is to make upload and download work similarly. Previously we only
required the source URL for uploads, but we used some logic + options
in the BlobDownloadOptions object to construct the destination URL for
downloads.

* Introduce StoragePathHelper to ease constructing destination URLs

* Use array extensions to ease locating transfers

* Add missing doc strings

* Update demo app to work with API changes

* Remove duplicate transferManager property

This was missed in the rebase of the last PR

* Teach downloader/uploader to speak only in URLs

The process of resolving blob+container to URL is now only ever done in
the client.

* Refactor TransferManager to use restorationId

This is a direct port of this work from @tjprescott's PR #238, just
broken out into its own commit.

* Modify the demo app to use the restorationId

* Standardize casing of `Url` in parameters and variables

* Construct upload/download URLs with template

This is to enable potential codegen in the future.

* Revise parameter names for upload/download to be more idiomatic

* Revise parameter names in TransferManager

* Move PathHelper inside StorageBlobClient

* Revert changes to examples

Now that the example project is compiling with these API changes, this
reverts all changes to the example project in order to make rebasing
easier once this PR is merged.