Keep a registry of restoration ID -> Client inside the TransferManager
This way, a customer only has to provide the restoration ID when
initializing a client. The client is later looked up from the registry
if it's found to be missing from the transfer. If no client exists in
the registry when the transfer is resumed, it will be marked as failed.