Local urls proposal (#258)
* Allow using any well-known directory as the base

Values that exist in `FileManager.SearchPathDirectory` use the string
"SearchPathDirectory.(rawValue):/" as the scheme (rawValue is a UInt). Values
that exist in the new `KnownDirectory` enum simply use their rawValue as the
scheme. The customer experience for either one is the same - you simply
call `url(for: .enumValue)`.

* Refactor PathHelper into LocalURL

We standardized things to operate on URLs, but then we discovered that
not all URLs are created equal. There are "real" URLs (instances of
`URL`) which can be used as-is, but then there are the special
"base-relative" URLs we came up with for representing local file paths,
that need to be stored in their "raw" form but resolved to a "real" URL
as needed.

Aside from reading documentation, it wasn't obvious to the customer that
there was a distinction - the public API accepts URLs for both source
and destination parameters. This commit refactors PathHelper into a new
structure called LocalURL, which represents a "base-relative" local
filesystem URL. Nothing changes about how we persist the local URLs to
CoreData, but the public API changes so that in places where the
parameter must be a special "base-relative" URL, it is now a LocalURL.

One advantage is that customers never have to be aware of the details of
how we represent the base-relative local URLs (via our custom URI
schemes). The customer can create and append to a LocalURL, and given a
LocalURL the customer can retrieve the real resolved `URL` it
represents.