Serialize LoggingPolicy logging
This makes sure that all log lines for a given request or response are
logged together. It doesn't prevent other threads from interrupting the
thread that's doing the logging though. Unfortunately there's no way to
guarantee that all logging operations in an entire app are done through
the same queue.