Add interactiveAuthPermitted to the request context
interactiveAuthPermitted will be added to the request context all the
time, EXCEPT when the downloader or uploader was previously nil and has
just been reconnected to the transfer. This means that all transfers
can potentially invoke an interactive authentication dialog when
started, but not when being auto-resumed by the transfer manager after
the app has restarted.