Overhaul SAS token credential implementation
We need to get a new SAS token for potentially every operation, since
tokens could be scoped to a single blob.

For performance, the customer will have to handle implementing a token
cache to avoid hitting their backend every single time even if it's
always returning the same token (e.g. an account- or container-level
token).