Cache SAS tokens by their requesting URL
This caches SAS tokens 1-1 with the URL that requests them. If the
provider returns an account- or container-level token, the cache is not
smart enough to know that the token can be used for all URLs under the
account / container. The customer would still need to implement a
caching strategy within their provider if they wanted to implement this
type of caching.